+++ studio/src/pages/RequestorPage/queries/hooks/useMakeProxiedRequest.ts
@@ -1,154 +1,13 @@
-import { PROBED_ROUTES_KEY } from "@/queries";
-import { validate } from "@scalar/openapi-parser";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { z } from "zod";
 import { useShallow } from "zustand/react/shallow";
-import { reduceFormDataParameters } from "./FormDataForm";
+import { reduceFormDataParameters } from "../../FormDataForm";
 import {
   type KeyValueParameter,
   reduceKeyValueParameters,
-} from "./KeyValueForm";
-import type { RequestorBody, RequestorResponseBody } from "./reducer/state";
-import { useRequestorStore } from "./store";
-import { RequestMethodSchema, RequestTypeSchema } from "./types";
-
-export const ProbedRouteSchema = z.object({
-  path: z.string(),
-  method: RequestMethodSchema,
-  handler: z.string(),
-  handlerType: z.enum(["route", "middleware"]),
-  currentlyRegistered: z.boolean(),
-  routeOrigin: z.enum(["discovered", "custom", "open_api"]),
-  openApiSpec: z.string().optional(),
-  requestType: RequestTypeSchema,
-  // NOTE - Added on the frontend, not stored in DB
-  isDraft: z.boolean().optional(),
-});
-
-export type ProbedRoute = z.infer<typeof ProbedRouteSchema>;
-
-const JsonSchema: z.ZodType<unknown> = z.lazy(() =>
-  z.union([
-    z.string(),
-    z.number(),
-    z.boolean(),
-    z.null(),
-    z.array(JsonSchema),
-    z.record(JsonSchema),
-  ]),
-);
-
-type JsonSchemaType = z.infer<typeof JsonSchema>;
-
-// TODO - Use validation schema
-export type Requestornator = {
-  app_requests: {
-    id: number;
-    requestUrl: string;
-    requestMethod: string;
-    requestRoute: string;
-    requestHeaders?: Record<string, string> | null;
-    requestQueryParams?: Record<string, string> | null;
-    requestPathParams?: Record<string, string> | null;
-    requestBody?: JsonSchemaType;
-    updatedAt: string;
-  };
-  // NOTE - can be undefined if request failed, at least that happened to me locally
-  app_responses: {
-    id: number;
-    responseStatusCode: string;
-    responseBody: string;
-    responseHeaders: Record<string, string>;
-    traceId: string;
-    isFailure: boolean;
-    failureReason: string | null;
-    updatedAt: string;
-  };
-};
-
-const REQUESTOR_REQUESTS_KEY = "requestorRequests";
-
-type ProbedRoutesResponse = {
-  baseUrl: string;
-  routes: ProbedRoute[];
-};
-
-function getProbedRoutes(): Promise<ProbedRoutesResponse> {
-  return fetch("/v0/app-routes").then((r) => r.json());
-}
-
-export function useProbedRoutes() {
-  return useQuery({
-    queryKey: [PROBED_ROUTES_KEY],
-    queryFn: getProbedRoutes,
-  });
-}
-
-export type Route = {
-  path: string;
-  method: string;
-  handler?: string;
-  handlerType?: "route" | "middleware";
-  routeOrigin?: "discovered" | "custom" | "open_api";
-  openApiSpec?: string;
-  requestType?: "http" | "websocket";
-  // NOTE - Added on the frontend, not stored in DB
-  isDraft?: boolean;
-};
-
-async function addRoutes(routes: Route | Route[]) {
-  return fetch("/v0/app-routes", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(routes),
-  }).then((r) => r.json());
-}
-
-export function useAddRoutes() {
-  const queryClient = useQueryClient();
-  const mutation = useMutation({
-    mutationFn: addRoutes,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [PROBED_ROUTES_KEY] });
-    },
-  });
-
-  return mutation;
-}
-
-function deleteRoute({
-  path,
-  method,
-}: {
-  path: string;
-  method: string;
-}) {
-  return fetch(`/v0/app-routes/${method}/${encodeURIComponent(path)}`, {
-    method: "DELETE",
-  }).then((r) => r.json());
-}
-
-export function useDeleteRoute() {
-  const queryClient = useQueryClient();
-  const mutation = useMutation({
-    mutationFn: deleteRoute,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [PROBED_ROUTES_KEY] });
-    },
-  });
-
-  return mutation;
-}
-const fetchQuery = () => fetch("/v0/all-requests").then((r) => r.json());
-
-export function useFetchRequestorRequests() {
-  return useQuery({
-    queryKey: [REQUESTOR_REQUESTS_KEY],
-    queryFn: fetchQuery,
-  });
-}
+} from "../../KeyValueForm";
+import type { RequestorBody, RequestorResponseBody } from "../../reducer";
+import { useRequestorStore } from "../../store";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { REQUESTOR_REQUESTS_KEY } from "./constants";
 
 export type MakeProxiedRequestQueryFn = ReturnType<
   typeof useMakeProxiedRequest
@@ -161,6 +20,7 @@ export function useMakeProxiedRequest() {
       setActiveResponse,
     })),
   );
+
   const queryClient = useQueryClient();
   const mutation = useMutation({
     mutationFn: makeProxiedRequest,
@@ -189,7 +49,7 @@ export function useMakeProxiedRequest() {
   return mutation;
 }
 
-export function makeProxiedRequest({
+function makeProxiedRequest({
   addServiceUrlIfBarePath,
   path,
   method,
@@ -387,18 +247,4 @@ function createBody(body: RequestorBody) {
 // NOTE - This is for urlencoded (not multipart)
 function createUrlEncodedBody(body: Record<string, string>) {
   return new URLSearchParams(body).toString();
-}
-
-export function useOpenApiParse(openApiSpec: string) {
-  const mutation = useMutation({
-    mutationFn: async (openApiSpec: string) => {
-      const { valid, schema } = await validate(openApiSpec);
-      if (!valid) {
-        throw new Error("Invalid OpenAPI spec");
-      }
-      return schema;
-    },
-    mutationKey: [openApiSpec],
-  });
-  return mutation;
 }
