+++ studio/src/pages/RequestorPage/queries/index.ts
@@ -1,0 +1,42 @@
+import { z } from "zod";
+
+const JsonSchema: z.ZodType<unknown> = z.lazy(() =>
+  z.union([
+    z.string(),
+    z.number(),
+    z.boolean(),
+    z.null(),
+    z.array(JsonSchema),
+    z.record(JsonSchema),
+  ]),
+);
+
+type JsonSchemaType = z.infer<typeof JsonSchema>;
+
+// TODO - Use validation schema
+export type Requestornator = {
+  app_requests: {
+    id: number;
+    requestUrl: string;
+    requestMethod: string;
+    requestRoute: string;
+    requestHeaders?: Record<string, string> | null;
+    requestQueryParams?: Record<string, string> | null;
+    requestPathParams?: Record<string, string> | null;
+    requestBody?: JsonSchemaType;
+    updatedAt: string;
+  };
+  // NOTE - can be undefined if request failed, at least that happened to me locally
+  app_responses: {
+    id: number;
+    responseStatusCode: string;
+    responseBody: string;
+    responseHeaders: Record<string, string>;
+    traceId: string;
+    isFailure: boolean;
+    failureReason: string | null;
+    updatedAt: string;
+  };
+};
+
+export * from "./hooks";
